+++ Libraries/RobsJuceModules/romos/Framework/romos_ModuleFactory.cpp
@@ -35,7 +35,7 @@ romos::Module* ModuleFactory::createModule(int id, const std::string& name, int 
   bool polyphonic)
 {
   ensureTypeInfoArrayAllocated();
-  rassert(id >= 0 && id < typeInfos->size());  // id out of range
+  rassert(id >= 0 && id < (int)typeInfos->size());  // id out of range
   // todo: if the id is out of range, return some kind of "Error" dummy module, maybe just a 
   // constant module outputting a 0, named "Error"? ..if that's possible
 
@@ -69,7 +69,7 @@ void ModuleFactory::deleteModule(romos::Module* moduleToDelete)
 int ModuleFactory::getModuleId(const std::string& fullTypeName)
 {
   ensureTypeInfoArrayAllocated();
-  for(int i = 0; i < typeInfos->size(); i++)
+  for(size_t i = 0; i < typeInfos->size(); i++)
     if((*typeInfos)[i]->fullName == fullTypeName)
       return i;
   return -1;
@@ -78,7 +78,7 @@ int ModuleFactory::getModuleId(const std::string& fullTypeName)
 ModuleTypeInfo* ModuleFactory::getModuleTypeInfo(const std::string& fullTypeName)
 {
   ensureTypeInfoArrayAllocated();
-  for(int i = 0; i < typeInfos->size(); i++)
+  for(size_t i = 0; i < typeInfos->size(); i++)
     if((*typeInfos)[i]->fullName == fullTypeName)
       return (*typeInfos)[i];
   return nullptr;
@@ -202,7 +202,7 @@ void ModuleFactory::registerPreBuiltContainers()
 void ModuleFactory::clearRegisteredTypes()
 {
   ensureTypeInfoArrayAllocated();
-  for(int i = 0; i < typeInfos->size(); i++)
+  for(size_t i = 0; i < typeInfos->size(); i++)
     delete (*typeInfos)[i];
   typeInfos->clear();
   delete typeInfos;
