+++ Libraries/RobsJuceModules/rosic/generators/rosic_TurtleSource.cpp
@@ -189,7 +189,7 @@ void TurtleSource::updateLineCommandIndices()
   //lineCommandIndices.resize(numLines);
   lineCommandIndices.resize(RAPT::rsMax(numLines,1));
   int j = 0;
-  for(int i = 0; i < turtleCommands.size(); i++) {
+  for(size_t i = 0; i < turtleCommands.size(); i++) {
     if(turtle.isLineCommand(turtleCommands[i])) {
       lineCommandIndices[j] = i;
       j++;
